# docs/swagger.yaml
definitions:
  dataapi.NewAgentProfileQuery:
    properties:
      agreeableness:
        type: number
      conscientiousness:
        type: number
      context_life_time_in_days:
        type: integer
      extraversion:
        type: number
      is_context_infinite:
        type: boolean
      neuroticism:
        type: number
      openness:
        type: number
    type: object
  dataapi.SearchAgentProfileQuery:
    properties:
      apikey:
        type: string
    type: object
  emotionapi.NewContextQuery:
    properties:
      api_key:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: zhanyunze0601@gmai.com
    name: 1Vewton
    url: https://github.com/1Vewton
  description: A service to simulate emotion for agent
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Emotion Simulator
  version: "1.0"
paths:
  /v1/data/addAgentProfile:
    post:
      consumes:
      - application/json
      description: Registers new emotion while returning a context id in the result
      parameters:
      - description: Query parameters
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/emotionapi.NewContextQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Registers new emotion while returning a context id in the result
      tags:
      - example
  /v1/data/getAgentProfile:
    post:
      consumes:
      - application/json
      description: Searches profile for agent
      parameters:
      - description: Query parameters
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dataapi.SearchAgentProfileQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Searches profile for agent
      tags:
      - example
  /v1/utils/health:
    get:
      consumes:
      - application/json
      description: check health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Checks health
      tags:
      - example
swagger: "2.0"